% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{plot.cv.hmlasso}
\alias{plot.cv.hmlasso}
\title{Plot a cross validation error path}
\usage{
\method{plot}{cv.hmlasso}(x, xlim = NULL, ylim = NULL, ...)
}
\arguments{
\item{x}{cross validated hmlasso model}

\item{xlim}{x range}

\item{ylim}{y range}

\item{...}{parameters of}
}
\description{
Plot a cross validation error path
}
\examples{
X_incompl <- as.matrix(iris[, 1:3])
X_incompl[1:5,1] <- NA
X_incompl[6:10,2] <- NA
y <- iris[, 4]
cv_fit <- cv.hmlasso(X_incompl, y, nlambda=50, lambda.min.ratio=1e-2)
plot(cv_fit)
plot(cv_fit$fit)

}
